Don't use company-clang backend
Not sure about this decision. Might revert later.

While working on a C project with TAGS file, it was better to use
completions from TAGS files which is given by company-etags
backend. But company-clang gets priority over  company-etags because
it is ahead in the list company-backends. So, I removed it
altogether. Also, it calls an external process for each completion. I
don't like this.